+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ## Forma 1.0.0
 
-* Added `subParser` and `value` combinators.
 * Changed signature of `mkFieldError` to accept `NonEmpty (SelectedName names)` 
   as first argument.
-* Added `runForm'` function that returns raw `BranchState`.
-* Exported `FieldError`'s value constructor.
+* Changed `runForm` function to return `BranchState`.
+* Added `toResponse` that converts `BranchState` to `Response`, which can be 
+  serialized to JSON.
+* Added `subParser` and `value` combinators.
+* Exported `showFieldPath` and `FieldError`'s value constructor.
 
 ## Forma 0.2.0